+++ src/Syntax.ts
@@ -6,7 +6,7 @@ import type { Either } from "@effect/data/Either"
 import type { LazyArg } from "@effect/data/Function"
 import type { Option } from "@effect/data/Option"
 import type { Predicate } from "@effect/data/Predicate"
-import * as internal from "@effect/parser/internal_effect_untraced/syntax"
+import * as internal from "@effect/parser/internal/syntax"
 import type { Parser } from "@effect/parser/Parser"
 import type { ParserError } from "@effect/parser/ParserError"
 import type { Printer } from "@effect/parser/Printer"
