+++ src/internal/bitset.ts
