+++ src/internal/common.ts
