+++ src/internal/lookupFunction.ts
@@ -7,8 +7,8 @@ import * as Number from "@effect/data/Number"
 import * as Option from "@effect/data/Option"
 import * as ReadonlyArray from "@effect/data/ReadonlyArray"
 import * as Cause from "@effect/io/Cause"
-import * as bitset from "@effect/parser/internal_effect_untraced/bitset"
-import * as common from "@effect/parser/internal_effect_untraced/common"
+import * as bitset from "@effect/parser/internal/bitset"
+import * as common from "@effect/parser/internal/common"
 import type * as Regex from "@effect/parser/Regex"
 
 const LookupFunctionSymbolKey = "@effect/parser/LookupFunction"
