+++ src/internal/parser.ts
@@ -4,10 +4,10 @@ import type { LazyArg } from "@effect/data/Function"
 import { dual, pipe } from "@effect/data/Function"
 import * as Option from "@effect/data/Option"
 import type { Predicate } from "@effect/data/Predicate"
-import * as recursive from "@effect/parser/internal_effect_untraced/parser/recursive"
-import * as stackSafe from "@effect/parser/internal_effect_untraced/parser/stack-safe"
-import * as parserError from "@effect/parser/internal_effect_untraced/parserError"
-import * as _regex from "@effect/parser/internal_effect_untraced/regex"
+import * as recursive from "@effect/parser/internal/parser/recursive"
+import * as stackSafe from "@effect/parser/internal/parser/stack-safe"
+import * as parserError from "@effect/parser/internal/parserError"
+import * as _regex from "@effect/parser/internal/regex"
 import type * as Parser from "@effect/parser/Parser"
 import type * as ParserError from "@effect/parser/ParserError"
 import type * as Regex from "@effect/parser/Regex"
