+++ src/internal/parser/recursive.ts
@@ -2,10 +2,10 @@ import * as Chunk from "@effect/data/Chunk"
 import * as Either from "@effect/data/Either"
 import * as List from "@effect/data/List"
 import * as Option from "@effect/data/Option"
-import * as common from "@effect/parser/internal_effect_untraced/common"
-import type * as parser from "@effect/parser/internal_effect_untraced/parser"
-import * as parserError from "@effect/parser/internal_effect_untraced/parserError"
-import * as _regex from "@effect/parser/internal_effect_untraced/regex"
+import * as common from "@effect/parser/internal/common"
+import type * as parser from "@effect/parser/internal/parser"
+import * as parserError from "@effect/parser/internal/parserError"
+import * as _regex from "@effect/parser/internal/regex"
 import type * as ParserError from "@effect/parser/ParserError"
 import type * as Regex from "@effect/parser/Regex"
