+++ src/internal/parser/stack-safe.ts
@@ -3,10 +3,10 @@ import * as Either from "@effect/data/Either"
 import type { LazyArg } from "@effect/data/Function"
 import * as List from "@effect/data/List"
 import * as Option from "@effect/data/Option"
-import * as common from "@effect/parser/internal_effect_untraced/common"
-import type * as parser from "@effect/parser/internal_effect_untraced/parser"
-import * as parserError from "@effect/parser/internal_effect_untraced/parserError"
-import * as regex from "@effect/parser/internal_effect_untraced/regex"
+import * as common from "@effect/parser/internal/common"
+import type * as parser from "@effect/parser/internal/parser"
+import * as parserError from "@effect/parser/internal/parserError"
+import * as regex from "@effect/parser/internal/regex"
 import type * as ParserError from "@effect/parser/ParserError"
 import type * as Regex from "@effect/parser/Regex"
