+++ src/internal/printer.ts
@@ -6,9 +6,9 @@ import { constVoid, dual, pipe } from "@effect/data/Function"
 import * as List from "@effect/data/List"
 import * as Option from "@effect/data/Option"
 import type { Predicate } from "@effect/data/Predicate"
-import * as chunkTarget from "@effect/parser/internal_effect_untraced/chunkTarget"
-import * as parserError from "@effect/parser/internal_effect_untraced/parserError"
-import * as _regex from "@effect/parser/internal_effect_untraced/regex"
+import * as chunkTarget from "@effect/parser/internal/chunkTarget"
+import * as parserError from "@effect/parser/internal/parserError"
+import * as _regex from "@effect/parser/internal/regex"
 import type * as ParserError from "@effect/parser/ParserError"
 import type * as Printer from "@effect/parser/Printer"
 import type * as Regex from "@effect/parser/Regex"
