+++ src/internal/syntax.ts
@@ -5,9 +5,9 @@ import { dual, pipe } from "@effect/data/Function"
 import * as Option from "@effect/data/Option"
 import type { Predicate } from "@effect/data/Predicate"
 import { tuple } from "@effect/data/Tuple"
-import * as _parser from "@effect/parser/internal_effect_untraced/parser"
-import * as _printer from "@effect/parser/internal_effect_untraced/printer"
-import * as _regex from "@effect/parser/internal_effect_untraced/regex"
+import * as _parser from "@effect/parser/internal/parser"
+import * as _printer from "@effect/parser/internal/printer"
+import * as _regex from "@effect/parser/internal/regex"
 import type * as Parser from "@effect/parser/Parser"
 import type * as ParserError from "@effect/parser/ParserError"
 import type * as Printer from "@effect/parser/Printer"
